Updated to latest libraries.
Former-commit-id: 6d6b2f1d994aae9d610cc37ce26757fdf78e5242

+++ DeviceCode/GHI/Libraries/GHI.Networking/Lib_GHI_Networking_PPPSerialModem.cpp
@@ -21,7 +21,7 @@ void PPPSerialModem::NativeEnable( CLR_RT_HeapBlock* pMngObj, UINT32 param0, HRE
     hr = CLR_E_NOT_SUPPORTED;
 }
 
-INT32 PPPSerialModem::NativeConnect( CLR_RT_HeapBlock* pMngObj, UINT32 param0, LPCSTR param1, LPCSTR param2, UINT32 param3, INT8 param4, HRESULT &hr )
+INT8 PPPSerialModem::NativeConnect( CLR_RT_HeapBlock* pMngObj, UINT32 param0, LPCSTR param1, LPCSTR param2, UINT32 param3, INT8 param4, HRESULT &hr )
 {
     hr = CLR_E_NOT_SUPPORTED;
